requirements.txt
pytest
biopython